requirements.txt
fastapi
flask
openpyxl
python-dotenv
openai
uvicorn
aiofiles
google-generativeai
mcp
pandas
httpx